Refactor angle detection and add ramp behavior
The game now properly sends the player flying if they are moving fast
enough to 'ramp' off the standard terrain (mainly RightTriangleTerrain.)
To accomodate this, the script 'game_pc_calc_shape_normal' was also
refactored, breaking down the calculations into individual scripts and
adjusting the use of the variable 'angle_resolution'. Furthemore,
DemoZone1 was updated with slightly different level design so it plays
more or less as it did originally.