Refactor terrain constructs (and add sinusoid)
The "Terrain Constructs" QuarterPipeConstruct and
QuarterEllipseConstruct were not working as intended; all relevant
code has been reworked to be easier to maintain. Auto-generated
terrain pieces are now instantiated in the Create Event, ensuring that
they are affected by DevVisualization. There are also new "constructs"
(ConcaveSinusoidConstruct and ConvexSinusoidConstruct) for users
to craft level geometry with.